pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "decker-pygame"
version = "0.5.0"
description = "A modern Python port of the classic 2001 Cyberspace RPG, Decker."
authors = [{ name = "Gemini Code Assist" }]
requires-python = ">=3.13"
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0.0",
    "pygame-ce>=2.5.0",
]

[project.scripts]
decker = "decker_pygame.presentation.main:main"

[project.optional-dependencies]
dev = [
    "commitizen",
    "mypy",
    "numpy",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "ruff",
]

[tool.ruff]
line-length = 88
exclude = [
    "typings/",
]

[tool.ruff.lint]
# E = pycodestyle errors, W = pycodestyle warnings, F = Pyflakes
# I = isort (import sorting)
# B = flake8-bugbear (finds likely bugs)
# C90 = mccabe (complexity checks)
# UP = pyupgrade (finds outdated syntax)
select = ["E", "W", "F", "I", "B", "C90", "UP"]
# Disable UP045 to prevent rewriting Optional[T] to T | None, which can cause
# issues with older libraries or mocking frameworks like unittest.mock.
ignore = ["UP045"]

[tool.ruff.format]
quote-style = "double"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_provider = "pep621"

[project.urls]
"Homepage" = "https://github.com/sdjaeb/decker-pygame"

[tool.mypy]
python_version = "3.13"
# Tell mypy where to find source code and custom type stubs
mypy_path = ["src", "typings"]
strict = true
explicit_package_bases = true
plugins = ["pydantic.mypy"]

# Pygame is a C-extension module, so we need to tell mypy to not worry about its internals
[[tool.mypy.overrides]]
module = "pygame.*"
ignore_missing_imports = true
ignore_errors = true
